third_party/WebKit/LayoutTests/fast/repaint/subpixel-shadow-included-in-invalidation-expected.txt

Tests whether a repaint rect encompasses box-shadow with subpixel offset when the shadow changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL Unexpected error: Uncaught TypeError: Cannot read property 'some' of undefined
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
